+++ Final Project/Common/src/main/java/commonproject/model/payment/Payment.java
@@ -19,5 +19,5 @@ public class Payment {
     @NotNull
     private int orderId;
 
-    private float fullPayment;
+    private BigDecimal fullPayment;
 }
